% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult.cox.mods.R
\name{adult.cox.mods}
\alias{adult.cox.mods}
\title{Conversion of adult capture data into the correct format for modeling seasonal (winter and summer) survival and adding covariates for cox proportional hazards modeling}
\usage{
adult.cox.mods(data, uni, uaidcol, startcol, mortcol, yearstart, yearend,
  seasons, winterstart, winterend, cause, plot, title)
}
\arguments{
\item{data}{data.frame that contains a column of unique animal identifier, start date, mortality date, and cause of mortality}

\item{uni}{vector of unique animal identifiers}

\item{startcol}{name of column with capture date}

\item{mortcol}{name of column that contains the date of mortality}

\item{yearstart}{year desired to begin modeling survival (e.g. beginning year of study)}

\item{yearend}{year desired to end modeling survival}

\item{seasons}{c('winter', 'summer')}

\item{winterstart}{month in numeric format that is desired to start winter (e.g. 11)}

\item{winterend}{month in numeric format that is desired for winter to end (e.g. 05)}

\item{cause}{vector of causes that require censoring (e.g. collar_failure, capturemort, etc.)}

\item{plot}{Logical. TRUE/FALSE. Whether a barplot of cumulative survival is desired}

\item{title}{desired title of plot}

\item{UAIDcol}{name of column where unique animals ids are stored}

\item{dateformat}{format of startcol and mortcols}
}
\value{
Returns a data.frame with animal ID, Year of survival, start date of modeling, end date of modeling, time alive (months), and status of animal (alive/censored = 0, dead = 1)
}
\description{
Convert capture and cause-specific mortality data of adults into the proper format for conducting survival analysis.This function will convert raw data
that contains start dates and mortality dates of individual animals and transform it into the proper format required by package survival
to model seasonal survival. Then, add covariates of choice for coxph models.
}
\keyword{adult,}
\keyword{analysis}
\keyword{kaplan-meier,}
\keyword{seasonal,}
\keyword{summer,}
\keyword{survival,}
\keyword{winter,}
