% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult.coxph.R
\name{adult.coxph}
\alias{adult.coxph}
\title{Create data with age and IFBFat to model survival}
\usage{
adult.coxph(data, uni, uaidcol, startcol, mortcol, yearstart, yearend,
  seasons, winterstart, winterend, cause, plot, title, grossweightcol,
  maxthickcol, bcscol, seasoncol)
}
\arguments{
\item{data}{database with age column in it}

\item{uni}{vector of unique ids in which survival should be calculated}

\item{uaidcol}{column where aids live}

\item{startcol}{name of column where capture date lives}

\item{mortcol}{name of column where mortality date lives}

\item{yearstart}{path to study area polygon (where you want to define availability)}

\item{yearend}{path to where predicted RF map should be written}

\item{seasons}{duh}

\item{winterstart}{duh}

\item{winterend}{duh}

\item{cause}{duh}

\item{plot}{duh}

\item{title}{duh}

\item{grossweightcol}{name of column where gross weight is stored}

\item{maxthickcol}{name of column where max thickness is stored}

\item{bcscol}{name of column where bcs is stored}

\item{seasoncol}{name of column where season (Fall/Spring) is located}
}
\value{
Returns a list object with RFData necessary to predict probably of use in RF models (elk, coyotes, and mountain lions)
}
\description{
Create data with age and IFBFat to model survival
}
\keyword{coyote,}
\keyword{elk,}
\keyword{extract,}
\keyword{forest,}
\keyword{lion,}
\keyword{mountain}
\keyword{random}
\keyword{raster,}
\keyword{sample}
