% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.mom.to.neo.R
\name{append.mom.to.neo}
\alias{append.mom.to.neo}
\title{bind mom aid to neonate capture data}
\usage{
append.mom.to.neo(neopath, mompath, pathout)
}
\arguments{
\item{neopath}{path to neonate database}

\item{mompath}{path to mom capture database}

\item{pathout}{path to write new neonate databse}
}
\value{
Returns a data.frame with original neonate capture database and mom's AID
}
\description{
bind mom aid to neonate capture database
}
\keyword{aid,}
\keyword{append}
\keyword{database,}
\keyword{neonate,}
