% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppendAID.R
\name{AppendAID}
\alias{AppendAID}
\title{Append Animal IDs created from CollarHistory function to original capture database}
\usage{
AppendAID(capdat, capturedate, mortalitydate, colhist, fileout)
}
\arguments{
\item{capdat}{path to capture database}

\item{capturedate}{name of column with capture date}

\item{mortalitydate}{name of column with mortality date}

\item{colhist}{data.frame created from CollarHistory function}

\item{fileout}{path to file and name of file where appended database shoudl be stored}
}
\value{
Returns a data.frame with original capture data.frame and a column appended with UAID
}
\description{
Append unique animals IDs created from CollarHistory function to capture database.
}
\examples{
\donttest{DBWithAID<-AppendAID(capdat = 'C:/Users/khuggler/Box Sync/DEER/Data/CleanDatabase/DB_WithAID_02082019.csv', capturedate = "CaptureDate", mortalitydate = "MortalityDate", 'dateformat = "\%m/\%d/\%Y', colhist = ColHist, fileout = 'C:/Users/khuggler/Box Sync/DEER/Data/CleanDatabase/DB_WithAID_02082019.csv')}
}
\keyword{ID,}
\keyword{adult,}
\keyword{animal}
\keyword{capture,}
\keyword{database}
