% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppendMR.R
\name{AppendMR}
\alias{AppendMR}
\title{Function to append movement rate categories to deer data}
\usage{
AppendMR(deerdata, elkdata, liondata, coyotedata)
}
\arguments{
\item{deerdata}{data.frame of deer gps data with repro info}

\item{elkdata}{data.frame of elk gps data with MRs}

\item{liondata}{data.frame of lion gps data with MRs}

\item{coyotedata}{data.frame of coyote gps data with MRs}
}
\value{
Returns a data.frame with all deer gps data with repro status and activity of lions, elk, and coyotes
}
\description{
Merge cleaned GPS data with the movement rates of elk, lions, and coyotes
}
\keyword{activity}
\keyword{deer,}
\keyword{gps,}
