% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.to.ssf.R
\name{bind.to.ssf}
\alias{bind.to.ssf}
\title{Function to append repro status and categorical activity rates back to extracted ssf data}
\usage{
bind.to.ssf(ssfdata, orig.data)
}
\arguments{
\item{ssfdata}{data.frame of ssf data (used and available)}

\item{orig.data}{original data.frame with categories of movement rates}
}
\value{
Returns a data.frame with all deer gps data needed for ssf analysis
}
\description{
Merge ssf data with original data
}
\keyword{activity}
\keyword{deer,}
\keyword{gps,}
