% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth.dates.R
\name{birth.dates}
\alias{birth.dates}
\title{append age to gps data}
\usage{
birth.dates(agepath, dbpath)
}
\arguments{
\item{agepath}{path to age table}

\item{dbpath}{path to capture database}
}
\value{
Returns original gps data with an age column
}
\description{
append ages of animals to gps data
}
\keyword{age,}
\keyword{capture,}
\keyword{gps}
