% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootFun.R
\name{boot.fun}
\alias{boot.fun}
\title{Boostrap random samples of data for RF models}
\usage{
boot.fun(data, sampsize, n.boot, mtry, cutoff, pred.names)
}
\arguments{
\item{sampsize}{proportion of data (unique Ids) to be taken for training data}

\item{n.boot}{number of bootstrapped samples to take}

\item{mtry}{desired mtry to test in RF models}

\item{cutoff}{parameter of cutoff to be used in RF models}

\item{pred.names}{names of predictor variables to be included in the model}

\item{data.frame}{of killsite data}
}
\value{
Returns a data.frame object with predictions of kill sites and available points
}
\description{
Build random forest models on bootstrapped samples of data
}
\keyword{forest}
\keyword{kill}
\keyword{lion,}
\keyword{mountain}
\keyword{prediction,}
\keyword{random}
\keyword{site,}
