% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildMcps.R
\name{BuildMcps}
\alias{BuildMcps}
\title{Function to build mcps for coyotes, lions, and elk}
\usage{
BuildMcps(deerdata, liondata, elkdata, coyotedata, percent)
}
\arguments{
\item{deerdata}{data.frame of deer gps data with repro info}

\item{liondata}{data.frame of lion gps data with MRs}

\item{elkdata}{data.frame of elk gps data with MRs}

\item{coyotedata}{data.frame of coyote gps data with MRs}

\item{percent}{percent of points to be contained in MCP}
}
\value{
Returns a data.frame with all deer gps data with repro status and activity of lions, elk, and coyotes
}
\description{
Build minimum convex polygons (with buffer if needed)
}
\keyword{activity}
\keyword{deer,}
\keyword{gps,}
