% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPredRiskModel.R
\name{BuildPredRiskModel}
\alias{BuildPredRiskModel}
\title{BuildRF Model with locations of mountain lion kill sites (summer and winter models)}
\usage{
BuildPredRiskModel(pred.data, ncpu, withold, method, combine.data)
}
\arguments{
\item{pred.data}{path of data to be used to build RF model}

\item{ncpu}{number of cpus to parallelize on}

\item{withold}{percent of data (decimal format) to withold for training model (default 0.25)}

\item{method}{method used for partitioning and training model ("bootstrap" or "repeatedcv")}

\item{raspath}{path to raster stack}

\item{studypath}{path to study area polygon (where you want to define availability)}

\item{pathout}{path to where predicted RF map should be written}
}
\value{
Returns a list object with RFData necessary to predict probably of use in RF models (elk, coyotes, and mountain lions)
}
\description{
Build RF Model with locations of mountain lion kill sites
}
\keyword{coyote,}
\keyword{elk,}
\keyword{extract,}
\keyword{forest,}
\keyword{lion,}
\keyword{mountain}
\keyword{random}
\keyword{raster,}
\keyword{sample}
