% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRFModels.R
\name{BuildRFModels}
\alias{BuildRFModels}
\title{Stack and extract all spatial data to elk, coyotes, and mountain lions (RF Encounter)}
\usage{
BuildRFModels(data, ncpu, withold, trees, pathout)
}
\arguments{
\item{data}{data to be used to build RF model}

\item{ncpu}{number of cpus to parallelize on}

\item{withold}{percent of data (decimal format) to withold for training model (default 0.25)}

\item{trees}{number of trees (maximum) to attempt to split on}

\item{pathout}{path to where predicted RF map should be written}

\item{studypath}{path to study area polygon (where you want to define availability)}
}
\value{
Returns a list object with RFData necessary to predict probably of use in RF models (elk, coyotes, and mountain lions)
}
\description{
Stack rasters and extract points data for elk, coyotes, and mountain lions
}
\keyword{coyote,}
\keyword{elk,}
\keyword{extract,}
\keyword{forest,}
\keyword{lion,}
\keyword{mountain}
\keyword{random}
\keyword{raster,}
\keyword{sample}
