% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.gps.R
\name{clean.gps}
\alias{clean.gps}
\title{Wrapper function with CombDat and CollarHistory that Merge Animal IDs and Sex, and Species to GPSData}
\usage{
clean.gps(atsfold, vecpath, capdat, spp)
}
\arguments{
\item{atsfold}{path to folder where ats data is located}

\item{vecpath}{path where vec data is located}

\item{capdat}{path to capture database}

\item{spp}{either "cervid" or "coyote"}
}
\value{
Returns a data.frame with all gps data, AnimalID, Sex, and Species
}
\description{
Merge raw GPS Data with Animal IDs, Sex, and Species
}
\keyword{ID,}
\keyword{animal}
\keyword{append}
\keyword{capture,}
\keyword{gps,}
