% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanGPSData.R
\name{CleanGPSData}
\alias{CleanGPSData}
\title{Merge Animal IDs and Sex, and Species to GPSData}
\usage{
CleanGPSData(gps, capdat, dateformat)
}
\arguments{
\item{gps}{gps data in the form of a data.frame or SpatialPointsDataFrame}

\item{capdat}{data.frame with all the capture information}

\item{dateformat}{format of TelemDate in gps data}
}
\value{
Returns a data.frame with all gps data, AnimalID, Sex, and Species
}
\description{
Merge raw GPS Data with Animal IDs, Sex, and Species
}
\examples{
\donttest{CleanData<-CleanGPSData(gps = yourgpsdata, capdat = yourcapturedata, dateformat = "\%Y-\%m-\%d")}
}
\keyword{ID,}
\keyword{animal}
\keyword{append}
\keyword{capture,}
\keyword{gps,}
