% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.dat.R
\name{climate.dat}
\alias{climate.dat}
\title{Read netCDF files, crop, and stack into climate raster}
\usage{
climate.dat(root.dir, study)
}
\arguments{
\item{root.dir}{path of root directory where spatial data is located}

\item{study}{path to study area shapefile}
}
\value{
Returns a list of raster bricks with all climate data
}
\description{
Read in netCDF files and manipulate to fit study area
}
\examples{

this code will read all of the files from the root directory (where netcdfs are stored), create the path names, read them in (raster function), pull out months needed, and then crop to study area and stack. 
}
\keyword{climate,}
\keyword{pdsi,}
\keyword{swe}
\keyword{temperature,}
