% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colhist.R
\name{colhist}
\alias{colhist}
\title{Conversion of adult capture data into capture history data}
\usage{
colhist(capdat)
}
\arguments{
\item{capdat}{data.frame with all the capture information}
}
\value{
Returns a data.frame with start and end date of every individual serial number along with an assigned animal ID)
}
\description{
Convert capture data into a record of every serial number and start and end dates of every individual captured
}
\examples{
\donttest{Hist<-CollarHistory(capdat = yourdata)}
}
\keyword{ID,}
\keyword{animal}
\keyword{capture,}
\keyword{history}
