% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ElkMR.R
\name{ElkMR}
\alias{ElkMR}
\title{Function to calculate movement rates of elk}
\usage{
ElkMR(gps, startdates, enddates, subspp, subsex)
}
\arguments{
\item{gps}{data.frame of gps data}

\item{startdates}{vector (date format) of the start dates to subset GPS data}

\item{enddates}{vector (date format) of the end dates to subset GPS data}

\item{subspp}{what species to subset}

\item{subsex}{what sex to subset}
}
\value{
Returns a data.frame with all gps data, AnimalID, Sex, and Species, and Reproductive Status (Repro or Non-Repro)
}
\description{
Merge cleaned GPS data with the movement rate of elk
}
\examples{
\donttest{ElkData<-ElkMR(gps = yourgpsdata, startdates = c('2017-05-01', '2018-05-01'), enddates = c('2017-09-01', '2018-09-01'), subspp = "ELK", subsex = "F")}
}
\keyword{gps,}
\keyword{neonate,}
\keyword{reproductive}
