% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFBFat.R
\name{IFBFat}
\alias{IFBFat}
\title{Stack and extract all spatial data to elk, coyotes, and mountain lions (RF Encounter)}
\usage{
IFBFat(x, weight = "lbs", GrossweightColName = "Gross.Weight..lb.",
  maxthickName = "MaxThickness", bcsName = "Body.Condition.Score",
  FetNum = 1.692, season = "Fall")
}
\arguments{
\item{x}{data to be used to build RF model}

\item{weight}{number of cpus to parallelize on}

\item{maxthickName}{number of trees (maximum) to attempt to split on}

\item{bcsName}{path to raster stack}

\item{FetNum}{path to study area polygon (where you want to define availability)}

\item{season}{path to where predicted RF map should be written}

\item{GrossewightColName}{percent of data (decimal format) to withold for training model (default 0.25)}
}
\value{
Returns a list object with RFData necessary to predict probably of use in RF models (elk, coyotes, and mountain lions)
}
\description{
Stack rasters and extract points data for elk, coyotes, and mountain lions
}
\keyword{coyote,}
\keyword{elk,}
\keyword{extract,}
\keyword{forest,}
\keyword{lion,}
\keyword{mountain}
\keyword{random}
\keyword{raster,}
\keyword{sample}
