% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv.hr.R
\name{indiv.hr}
\alias{indiv.hr}
\title{Build kernel density home ranges of individual animals}
\usage{
indiv.hr(gps)
}
\arguments{
\item{gps}{SpatialPointsDataFrame that has an 'id' column, and 'x' and 'y' column only. 'id' column should be named according to how home ranges should be built}

\item{reference}{bandwidth type can be either "href" or "LSCV"}

\item{bandmethod}{method for choosing correct bandwidth options are "none" to stick with reference of LSCV bandwidth, or "adhoc" to use an ad-hoc smooth parameter. If ad hoc is chosen, reference bandwidth is reduced by 10% increments and all home ranges are stored in the directory}

\item{percent}{percent volume to be calculated and size of home range to be saved}

\item{unout}{units for home range to be read out (see getverticeshr for options)}

\item{dirout}{path to folder where home ranges should be saved}

\item{season}{name of column where season (Fall/Spring) is stored}
}
\value{
Returns a original dataframe with appended columns of IFBFat and Scaled IFBFat
}
\description{
Build kernel home ranges for individual animals and produce vertices
}
\keyword{IFBFat,}
\keyword{body}
\keyword{condition,}
\keyword{max}
\keyword{thicknes,}
\keyword{weight}
