% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredRiskData.R
\name{PredRiskData}
\alias{PredRiskData}
\title{BuildRF model of predation risk}
\usage{
PredRiskData(clustpath, dbpath, neopath, subset, startdates, enddates,
  raspath, studypath, nsamps, pathout)
}
\arguments{
\item{clustpath}{path to cluster data}

\item{dbpath}{path to adult capture database}

\item{neopath}{path to neoante database}

\item{subset}{Logical TRUE/FALSE. Whether or not data should be subset to a certaind date range}

\item{startdates}{vector of desired start dates}

\item{enddates}{vector of desired end dates}

\item{raspath}{path to raster stack}

\item{studypath}{path to study area polygon (where you want to define availability)}

\item{nsamps}{number of random samples per 1 used point}

\item{pathout}{path to write csv of predation risk data}
}
\value{
Returns a data.frame object with RFData necessary to predict probability of kill in RF models mountain lions only
}
\description{
Build random forest model to predict probability of kill
}
\keyword{deer}
\keyword{lion,}
\keyword{mountain}
\keyword{predation}
\keyword{risk,}
