% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessSpatial.R
\name{ProcessSpatial}
\alias{ProcessSpatial}
\title{Process all spatial data for analyses}
\usage{
ProcessSpatial(raspath, sagepath)
}
\arguments{
\item{raspath}{path to rasters already clipped to study area}

\item{sagepath}{path to location of sage rasters}

\item{studypath}{path to location of study area polygon}
}
\value{
Returns a raster stack with all spatial data cropped to large study area
}
\description{
Clip, re-project, and stack rasters for analyes
}
\keyword{data,}
\keyword{raster,}
\keyword{spatial}
\keyword{stack}
