% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFEncounterData.R
\name{RFEncounterData}
\alias{RFEncounterData}
\title{Stack and extract all spatial data to elk, coyotes, and mountain lions (RF Encounter)}
\usage{
RFEncounterData(elkdata, liondata, coyotedata, raspath, studypath)
}
\arguments{
\item{elkdata}{object of data.frame of elk GPS data}

\item{liondata}{object of data.frame of lion GPS data}

\item{coyotedata}{object of data.frame of coyote GPS data}

\item{raspath}{path to raster stack}

\item{studypath}{path to study area polygon (where you want to define availability)}
}
\value{
Returns a list object with RFData necessary to predict probably of use in RF models (elk, coyotes, and mountain lions)
}
\description{
Stack rasters and extract points data for elk, coyotes, and mountain lions
}
\keyword{coyote,}
\keyword{elk,}
\keyword{extract,}
\keyword{forest,}
\keyword{lion,}
\keyword{mountain}
\keyword{random}
\keyword{raster,}
\keyword{sample}
