% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleFun.R
\name{SampleFun}
\alias{SampleFun}
\title{Sample Fun}
\usage{
SampleFun(raspath, nmin, nmax, by, poly, proj)
}
\arguments{
\item{raspath}{path to zip folder of rasters}

\item{nmin}{number of minimum available points to test}

\item{nmax}{number of maximum available points to test}

\item{by}{interval of number of points to try (10, 20, 100, etc.)}

\item{poly}{path to study area polygon}

\item{proj}{projection (character) of study area polygon}
}
\value{
Returns a data.frame object with mean extracted values of each raster at each of the specified point intervals
}
\description{
Identify the number of sample points necessary to adequately represent variation in study area
}
\keyword{available}
\keyword{points,}
\keyword{rsf}
\keyword{sampling,}
