% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleveggies.R
\name{sampleveggies}
\alias{sampleveggies}
\title{Sample random points from neonates for vegetation work}
\usage{
sampleveggies(x, a, year)
}
\arguments{
\item{x}{path to neonate data}

\item{a}{list of time intervals to sample (in days)}

\item{year}{year of interest to subset neonate data}

\item{x}{path to neonate data}

\item{a}{list of time intervals to sample (in days)}

\item{year}{year of interest to subset neonate data}
}
\description{
Sample random points from neonates for vegetation work

Sample random points from neonates for vegetation work
}
\keyword{gps,}
\keyword{neonate}
\keyword{sample,}
\keyword{vegetation,}
