% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VegSampleFun.R
\name{VegSampleFun}
\alias{VegSampleFun}
\title{Sample random points from neonates for vegetation work}
\usage{
VegSampleFun(x, a, year)
}
\arguments{
\item{x}{path to neonate data}

\item{a}{list of time intervals to sample (in days)}

\item{year}{year of interest to subset neonate data}
}
\value{
returns a csv file with all neonate sample data
}
\description{
Sample random points from neonates for vegetation work
}
\keyword{gps,}
\keyword{neonate}
\keyword{sample,}
\keyword{vegetation,}
