% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yote.gps.R
\name{yote.gps}
\alias{yote.gps}
\title{Append AID to GPS data and subset to desired time period}
\usage{
yote.gps(yotedat, gps)
}
\arguments{
\item{yotedat}{path to coyote capture database}

\item{gps}{object identifying gps data file}
}
\value{
Returns a data.frame with all gps data, AnimalID, Sex,
}
\description{
Merge raw GPS Data with Animal IDs, and Sex
}
\keyword{animal}
\keyword{coyote,}
\keyword{id,}
\keyword{movement}
\keyword{rate}
