% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult.coxph.R
\name{adult.cox.mods}
\alias{adult.cox.mods}
\title{Stack and extract all spatial data to elk, coyotes, and mountain lions (RF Encounter)}
\usage{
adult.cox.mods(data, uni, uaidcol, startcol, mortcol, yearstart, yearend,
  seasons, winterstart, winterend, cause, plot, title)
}
\arguments{
\item{data}{to be used to build RF model}

\item{uni}{number of cpus to parallelize on}

\item{uaidcol}{percent of data (decimal format) to withold for training model (default 0.25)}

\item{startcol}{number of trees (maximum) to attempt to split on}

\item{mortcol}{path to raster stack}

\item{yearstart}{path to study area polygon (where you want to define availability)}

\item{yearend}{path to where predicted RF map should be written}

\item{seasons}{duh}

\item{winterstart}{duh}

\item{winterend}{duh}

\item{cause}{duh}

\item{plot}{duh}

\item{title}{duh}
}
\value{
Returns a list object with RFData necessary to predict probably of use in RF models (elk, coyotes, and mountain lions)
}
\description{
Stack rasters and extract points data for elk, coyotes, and mountain lions
}
\keyword{coyote,}
\keyword{elk,}
\keyword{extract,}
\keyword{forest,}
\keyword{lion,}
\keyword{mountain}
\keyword{random}
\keyword{raster,}
\keyword{sample}
