% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdultAnnualSurv.R
\name{AdultAnnualSurv}
\alias{AdultAnnualSurv}
\title{Conversion of adult capture data into the correct format for modeling annual survival}
\usage{
AdultAnnualSurv(data, format, uaidcol, capcol, mortcol, yearstart, yearend,
  cause)
}
\arguments{
\item{data}{data.frame that contains a column of unique animal identifier, start date, mortality date, and cause of mortality}

\item{format}{format of mortcol}

\item{uaidcol}{name of column where unique animal identifiers are located}

\item{capcol}{name of column where capture date is stored}

\item{mortcol}{name of column that contains the date of mortality}

\item{yearstart}{year desired to begin modeling survival (e.g. beginning year of study)}

\item{yearend}{year desired to end modeling survival}

\item{cause}{vector of causes that require censoring (e.g. collar_failure, capturemort, etc.)}
}
\value{
Returns a data.frame with animal ID, start date of modeling, end date of modeling, status of animal (alive = 0, dead = 1), and number of months alive during time period
}
\description{
Convert capture and cause-specific mortality data of adults into the proper format for conducting survival analysis.This function will convert raw data
that contains start dates and mortality dates of individual animals and transform it into the proper format required by package survival
to model annual survival.
}
\examples{
\donttest{AdultSurv<-AdultAnnualSurv(data = yourdata, uni = uniquevector, mortcol = "MortalityDate", yearstart = 2015, yearend = 2019 , cause = "CaptureMort")}

}
\keyword{adult,}
\keyword{analysis}
\keyword{annual,}
\keyword{kaplan-meier,}
\keyword{survival,}
