% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.age.R
\name{append.age}
\alias{append.age}
\title{append ages and calculate new age for every capture event}
\usage{
append.age(agepath, cappath, pathout)
}
\arguments{
\item{agepath}{path to where age table lives}

\item{cappath}{path to where capture data lives}

\item{pathout}{path to where new data.frame should be written}
}
\value{
Returns original capture data.frame with age column of each animal
}
\description{
append ages of animals to databaes and calculate age for every capture event
}
\keyword{age,}
\keyword{capture}
