% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.age.R
\name{append.age}
\alias{append.age}
\title{append age to gps data}
\usage{
append.age(agepath, cappath, idcol, gps, type)
}
\arguments{
\item{cappath}{path to where capture data (with birth dates) lives}

\item{idcol}{name of column where animal ID exists}

\item{gps}{name of gps dataframe}

\item{type}{type of append to do. Either "gps" or "database"}
}
\value{
Returns original gps data with an age column
}
\description{
append ages of animals to gps data
}
\keyword{age,}
\keyword{capture,}
\keyword{gps}
