% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.gps.R
\name{clean.gps}
\alias{clean.gps}
\title{Wrapper function with CombDat and CollarHistory that Merge Animal IDs and Sex, and Species to GPSData}
\usage{
clean.gps(vecdata, vecpath, usernames, passwords, tempdir, ST, STUser,
  STPass, cType, capdat, dateformat)
}
\arguments{
\item{vecdata}{Logical. TRUE/FALSE. Whether vecdata is needed to be downloaded}

\item{vecpath}{path where vec data is located}

\item{usernames}{character vector of usernames to download}

\item{passwords}{character vector of passwords in same order as usernames}

\item{tempdir}{temporary directory for data to download (can be desktop)}

\item{ST}{Logical. TRUE/FALSE. Whether you need SirTrack data}

\item{STUser}{SirTrack Username}

\item{STPass}{SirTrack Password}

\item{cType}{"ATS/IRID" or "ATS/GSTAR"}

\item{capdat}{data.frame with all the capture information}

\item{dateformat}{format of TelemDate in gps data}
}
\value{
Returns a data.frame with all gps data, AnimalID, Sex, and Species
}
\description{
Merge raw GPS Data with Animal IDs, Sex, and Species
}
\examples{
\donttest{CleanData<-CleanGPSData(gps = yourgpsdata, capdat = yourcapturedata, dateformat = \%Y-\%m-\%d)}
}
\keyword{ID,}
\keyword{animal}
\keyword{append}
\keyword{capture,}
\keyword{gps,}
