% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleangps.R
\name{cleangps}
\alias{cleangps}
\title{Wrapper function with CombDat and CollarHistory that Merge Animal IDs and Sex, and Species to GPSData}
\usage{
cleangps(
  atsfold,
  veckeys,
  sirtrack,
  gstar,
  capdat,
  spp,
  buffer.capture,
  buffer
)
}
\arguments{
\item{atsfold}{path to folder where ats data is located}

\item{veckeys}{path to where vectronics keys are located}

\item{sirtrack}{path to where sirtrack data is located}

\item{gstar}{path to where gstar data is located}

\item{capdat}{path to capture database}

\item{spp}{either "cervid" or "coyote"}

\item{buffer.capture}{Logical. True if you would like to remove locations surrounding captures (user defined buffer in days) False if you would like to include captures in gps data}

\item{buffer}{numeric number of days to buffer around captures}
}
\value{
Returns a data.frame with all gps data, AnimalID, Sex, and Species
}
\description{
Merge raw GPS Data with Animal IDs, Sex, and Species
}
\examples{
=============================
read in ATS data
=============================

==========================================
read in vectronics data and bind together
=========================================

======================================
merge vectronics and ats data
======================================
=============================
create collar history table
=============================
}
\keyword{ID,}
\keyword{animal}
\keyword{append}
\keyword{capture,}
\keyword{gps,}
