% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFBFatFun.R
\name{IFBFat}
\alias{IFBFat}
\title{Conversion of adult max thickness and body condition score into IFBFat metric}
\usage{
IFBFat(x, weight = "lbs", GrossweightColName = "Gross.Weight..lb.",
  maxthickName = "MaxThickness", bcsName = "Body.Condition.Score",
  FetNum = 1.692, season = "Fall")
}
\arguments{
\item{x}{data.frame of capture data}

\item{weight}{'lbs'}

\item{GrossweightColName}{name of column where gross weight measurement is stored}

\item{maxthickName}{name of column where max thickness is stored}

\item{bcsName}{name of column where body condition score is stored}

\item{FetNum}{1.692}

\item{Season}{a column indicating whether capture event is in the "Fall" or "Spring"}
}
\value{
Returns original capture data.frame with appending columns of IFBFat metrics
}
\description{
Convert measurements of max thickness and body condition score into IFBFat metric
}
\keyword{IFBFat,}
\keyword{adult,}
\keyword{body}
\keyword{condition}
\keyword{score}
\keyword{seasonal,}
\keyword{summer,}
\keyword{weight,}
\keyword{winter,}
