% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFBFat.R
\name{IFBFat}
\alias{IFBFat}
\title{Calculate IFBFat}
\usage{
IFBFat(x, weight = "lbs", GrossweightColName = "Gross.Weight..lb.",
  maxthickName = "MaxThickness", bcsName = "Body.Condition.Score",
  FetNum = 1.692, season = "Fall")
}
\arguments{
\item{x}{capture database}

\item{weight}{'lbs'}

\item{maxthickName}{name of column where max thickness is stored}

\item{bcsName}{name of column where body condition score is stored}

\item{FetNum}{path to study area polygon (where you want to define availability)}

\item{season}{name of column where season (Fall/Spring) is stored}

\item{GrossewightColName}{name of column where gross weight is stored}
}
\value{
Returns a original dataframe with appended columns of IFBFat and Scaled IFBFat
}
\description{
Calculate IFBFat from weight, max thickness, body condition score, and fetal number.
}
\keyword{IFBFat,}
\keyword{body}
\keyword{condition,}
\keyword{max}
\keyword{thicknes,}
\keyword{weight}
