% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssf.fun.R
\name{ssf.fun}
\alias{ssf.fun}
\title{build ssf with deer data}
\usage{
ssf.fun(dat, datecol, idcol, plot, pathout, nsamps)
}
\arguments{
\item{dat}{deer data}

\item{datecol}{name of column where telemdate is stored}

\item{idcol}{name of column where aid is stored}

\item{plot}{TRUE/FALSE whether plots should be generated}

\item{pathout}{path to where ssf data should be written}

\item{nsamps}{number of random points for every step}
}
\value{
Returns a data.frame object with all sampled ssf data
}
\description{
sample available points from distribution for ssf data
}
\keyword{distribution}
\keyword{sample,}
\keyword{selection,}
\keyword{ssf,}
\keyword{step}
